+++ lib/Rename/Model/ReferenceFinder/PredefinedReferenceFinder.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phpactor\Rename;
+namespace Phpactor\Rename\Model\ReferenceFinder;
 
 use Phpactor\ReferenceFinder\PotentialLocation;
 use Phpactor\ReferenceFinder\ReferenceFinder;
@@ -8,8 +8,11 @@ use Phpactor\TextDocument\ByteOffset;
 use Phpactor\TextDocument\TextDocument;
 use Generator;
 
-class Model implements ReferenceFinder
+class PredefinedReferenceFinder implements ReferenceFinder
 {
+    /**
+     * @var PotentialLocation[]
+     */
     private array $locations;
 
     public function __construct(PotentialLocation ...$locations)
