+++ lib/Rename/Tests/Adapter/ReferenceFinder/ReferenceRenamerIntegrationTestCase.php
@@ -3,8 +3,9 @@
 namespace Phpactor\Rename\Tests\Adapter\ReferenceFinder;
 
 use PHPUnit\Framework\TestCase;
-use Phpactor\Rename\Model;
+use Phpactor\ReferenceFinder\ReferenceFinder;
 use Phpactor\ReferenceFinder\PotentialLocation;
+use Phpactor\Rename\Model\ReferenceFinder\PredefinedReferenceFinder;
 use Phpactor\TextDocument\ByteOffset;
 use Phpactor\TextDocument\Location;
 use Phpactor\TextDocument\TextDocument;
@@ -12,11 +13,11 @@ use Phpactor\TextDocument\TextDocument;
 abstract class ReferenceRenamerIntegrationTestCase extends TestCase
 {
     /**
-     * @return ByteOffset[]
+     * @param ByteOffset[] $references
      */
-    public function offsetsToReferenceFinder(TextDocument $textDocument, array $references): Model
+    public function offsetsToReferenceFinder(TextDocument $textDocument, array $references): ReferenceFinder
     {
-        return new Model(...array_map(function (ByteOffset $reference) use ($textDocument) {
+        return new PredefinedReferenceFinder(...array_map(function (ByteOffset $reference) use ($textDocument) {
             return PotentialLocation::surely(new Location($textDocument->uri(), $reference));
         }, $references));
     }
